+++ src/BackgroundThrust/Obsoleted/BackgroundSolverEngine.cs
@@ -1,0 +1,6 @@
+using System;
+
+namespace BackgroundThrust.Obsoleted;
+
+[Obsolete("BackgroundSolverEngine is no longer needed and can be replaced by BackgroundEngine")]
+public class BackgroundSolverEngine : BackgroundEngine { }